fix: widen current_offset to uint64_t so RVA overflow guard works, zero stack_start_out on alloc failure
- current_offset was uint32_t, making the > UINT32_MAX check always
  false. Widening to uint64_t lets the guard actually detect overflow.
- write_thread_stack now zeros *stack_start_out on malloc failure,
  matching the macOS counterpart.

+++ src/backends/native/minidump/sentry_minidump_common.h
@@ -11,7 +11,7 @@
  */
 typedef struct {
     int fd;
-    uint32_t current_offset;
+    uint64_t current_offset;
 } minidump_writer_base_t;
 
 /**
